refactor db util
