+++ app/src/main/java/roide/nanod/popularmovies/network/MovieAPI.java
@@ -14,27 +14,6 @@ public interface MovieAPI
     String QUERY_SORT_BY = "sort_by";
     String QUERY_PAGE = "page";
 
-    String SORT_HIGHEST_RATED = "vote_average.desc";
-    String SORT_MOST_POPULAR = "popularity.desc";
-
-    enum SortOrder
-    {
-        HIGHEST_RATED("vote_average.desc"),
-        MOST_POPULAR("popularity.desc");
-
-        private final String mSortOrder;
-        private SortOrder(String sortOrder)
-        {
-            mSortOrder = sortOrder;
-        }
-
-        @Override
-        public String toString()
-        {
-            return mSortOrder;
-        }
-    }
-
     @GET("/discover/movie")
     void discoverMovies(
             @Query(QUERY_API_KEY) String apiKey,
